+++ processor/src/test/fixtures/input/com/example/contribute_to/ContributeToOnActAsStingComponentModel.java
@@ -1,0 +1,18 @@
+package com.example.contribute_to;
+
+import sting.ActAsStingComponent;
+import sting.ContributeTo;
+
+public class ContributeToOnActAsStingComponentModel
+{
+  @ActAsStingComponent
+  @interface MyFramework
+  {
+  }
+
+  @MyFramework
+  @ContributeTo( "Foo" )
+  static class MyComponentType
+  {
+  }
+}