+++ processor/src/test/fixtures/input/com/example/named/NamedOnActAsStingComponentModel.java
@@ -1,0 +1,18 @@
+package com.example.named;
+
+import sting.ActAsStingComponent;
+import sting.Named;
+
+public class NamedOnActAsStingComponentModel
+{
+  @ActAsStingComponent
+  @interface MyFramework
+  {
+  }
+
+  @MyFramework
+  @Named( "Foo" )
+  static class MyComponentType
+  {
+  }
+}