+++ processor/src/test/fixtures/input/com/example/named/NamedOnCtorParamInActAsStingComponentModel.java
@@ -1,0 +1,20 @@
+package com.example.named;
+
+import sting.ActAsStingComponent;
+import sting.Named;
+
+public class NamedOnCtorParamInActAsStingComponentModel
+{
+  @ActAsStingComponent
+  @interface MyFramework
+  {
+  }
+
+  @MyFramework
+  static class MyComponentType
+  {
+    MyComponentType( @Named( "Foo" ) int i )
+    {
+    }
+  }
+}